+++ io.asgardio.java.oidc.sdk/src/main/java/io/asgardio/java/oidc/sdk/DefaultOIDCManager.java
@@ -43,10 +43,10 @@ import com.nimbusds.oauth2.sdk.token.AccessToken;
 import com.nimbusds.oauth2.sdk.token.RefreshToken;
 import com.nimbusds.openid.connect.sdk.Nonce;
 import com.nimbusds.openid.connect.sdk.claims.IDTokenClaimsSet;
-import io.asgardio.java.oidc.sdk.bean.LogoutRequest;
+import io.asgardio.java.oidc.sdk.request.model.LogoutRequest;
 import io.asgardio.java.oidc.sdk.bean.RequestContext;
 import io.asgardio.java.oidc.sdk.bean.SessionContext;
-import io.asgardio.java.oidc.sdk.bean.AuthenticationRequest;
+import io.asgardio.java.oidc.sdk.request.model.AuthenticationRequest;
 import io.asgardio.java.oidc.sdk.bean.User;
 import io.asgardio.java.oidc.sdk.config.model.OIDCAgentConfig;
 import io.asgardio.java.oidc.sdk.exception.SSOAgentClientException;
