+++ io.asgardio.java.oidc.sdk/src/test/java/io/asgardio/java/oidc/sdk/request/OIDCRequestBuilderTest.java
@@ -22,8 +22,8 @@ import com.nimbusds.jwt.JWT;
 import com.nimbusds.jwt.JWTParser;
 import com.nimbusds.oauth2.sdk.Scope;
 import com.nimbusds.oauth2.sdk.id.ClientID;
-import io.asgardio.java.oidc.sdk.bean.AuthenticationRequest;
-import io.asgardio.java.oidc.sdk.bean.LogoutRequest;
+import io.asgardio.java.oidc.sdk.request.model.AuthenticationRequest;
+import io.asgardio.java.oidc.sdk.request.model.LogoutRequest;
 import io.asgardio.java.oidc.sdk.bean.RequestContext;
 import io.asgardio.java.oidc.sdk.bean.SessionContext;
 import io.asgardio.java.oidc.sdk.config.model.OIDCAgentConfig;
